// src/werp/enums/SessionKeys.ts
enum SessionKeys {
    SIGN_OUT_ALREADY_NOTIFICATION = 'SIGN_OUT_ALREADY_NOTIFICATION',
    SIGN_IN_NOTIFICATION = 'SIGN_IN_NOTIFICATION',
    OFF_WORK_NOTIFICATION = 'OFF_WORK_NOTIFICATION',
    SIGN_OUT_NOTIFICATION = 'SIGN_OUT_NOTIFICATION',
    ATTENDANCE_TIMER = 'ATTENDANCE_TIMER',
    COMPANY_EMPLOYEE_COUNT = 'COMPANY_EMPLOYEE_COUNT',
}

export default SessionKeys;